Enhance clickable area and style of the intro link in MapMenu
Improved the user interface for the introduction link in the MapMenu component.
Added padding and a pointer cursor style to make the link more accessible and
visually indicative of its interactivity. These changes aim to enhance the user
experience by making the introduction link more noticeable and easier to click.